requirements.txt
agno==1.7.9
aiohttp
chromadb==0.5.15
faiss-cpu
fastapi==0.116.1
PyMuPDF
groq==0.31.0
httpx==0.28.1
google-genai
langdetect==1.0.9
rank_bm25
nltk==3.8.1
numpy
ruff
openai==1.99.5
pandas==2.3.1
Pillow==11.3.0
protobuf==6.31.1
pydantic
pydantic_core
pydantic_settings
pytesseract==0.3.13
python-dotenv==1.1.1
python_docx==1.2.0
python_pptx==1.0.2
rank_bm25==0.2.2
scikit_learn
sentence_transformers==4.1.0
starlette==0.47.2
tiktoken==0.9.0
torch==2.5.1
uvicorn

